I forgot the reflection
This project went pretty smoothly because I have some experience with maya so this basic modeling was easy, I just had to refresh myself on the interface. When I was making the model I wanted to make a funky modern vibe because I felt like just doing a basic chair and table was boring.