Update Room_simple.ma
It took me some time to figure out how to freeze transformations and how to delete history. Attempted to choose what to model to awhile. I create the vase by using a cylinder and adding subdivisions. I selected the faces to create the shape of the vase and did some shaping by manipulating  the sides with the edge tool. To make the bamboo I made another cylinder and adding subdivisions. I added edge loops next to the subdivisions and extrude the center edge all the way around. I then duplicates until the vase was full with some mild changes in the bamboos shape.